refactor(lotto): RandomLottoFactory 를 삭제하고 LottoFactory 를 클래스로 변경
