+++ src/main/kotlin/lotto/domain/LottoFactory.kt
@@ -1,13 +1,13 @@
-package lotto.domain.factory
+package lotto.domain
 
 import lotto.constant.LOTTO_MAXIMUM_NUMBER
 import lotto.constant.LOTTO_MINIMUM_NUMBER
 import lotto.constant.LOTTO_SIZE
-import lotto.domain.Lotto
-import lotto.domain.LottoNumber
 
-class RandomLottoFactory : LottoFactory {
-    override fun createLotto(): Lotto = Lotto(getRandomNumbers().map { LottoNumber.from(it) })
+class LottoFactory {
+
+    fun createLotto(numbers: List<Int> = getRandomNumbers()): Lotto =
+        Lotto(numbers.map { number -> LottoNumber.from(number) })
 
     private fun getRandomNumbers(): List<Int> = (LOTTO_MINIMUM_NUMBER..LOTTO_MAXIMUM_NUMBER).shuffled().take(LOTTO_SIZE)
 }
