--- src/main/kotlin/lotto/domain/factory/LottoFactory.kt
@@ -1,7 +1,0 @@
-package lotto.domain.factory
-
-import lotto.domain.Lotto
-
-interface LottoFactory {
-    fun createLotto(): Lotto
-}
